PP management no longer shows epic mods if the actor does not have the edge
