+++ scripts/brsw2-config.js
@@ -95,6 +95,12 @@ export const UNTRAINED_SKILLS = [
   "(unskilled)",
 ];
 
+export const ARCANE_MASTERY_EDGES = [
+  "BRSW.EdgeName.ArcaneMastery",
+  "BRSW.EdgeName.DivineMastery",
+  "BRSW.EdgeName.Epic-Mastery",
+];
+
 /**
 / Translation map for attributes
 */
