+++ scripts/cards_common.js
@@ -43,11 +43,6 @@ export const BRSW_CONST = {
   TYPE_UNSHAKE_CARD: 13,
   TYPE_UNSTUN_CARD: 14,
   TYPE_RESULT_CARD: 100,
-  ARCANE_MASTERY_EDGES: [
-    "BRSW.EdgeName.ArcaneMastery",
-    "BRSW.EdgeName.DivineMastery",
-    "BRSW.EdgeName.Epic-Mastery",
-  ],
 };
 
 /**
